fix a bug, when label height < text height, label will blank

+++ TYAttributedLabelDemo/TYAttributedLabel/TYTextContainer.h
@@ -63,7 +63,7 @@
 /**
  *  获取文本高度
  */
-- (int)getHeightWithFramesetter:(CTFramesetterRef)framesetter width:(CGFloat)width;
+- (CGFloat)getHeightWithFramesetter:(CTFramesetterRef)framesetter width:(CGFloat)width;
 
 @end
